refactor: move normal user task migration to ProcessInstanceMigrationUserTaskBehavior
-move task migration to the relevant behavior class
-fix linting errors
